+++ CHANGELOG.md
@@ -19,4 +19,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.2.0] - 2020-12-15
 ### Added
-- Feature: Add cli command `hello-world` for printing hello world message with emoji 
+- Feature: Add cli command `hello-world` for printing hello world message with emoji
+
+## [v0.3.0] - 2021-01-17
+### Added
+- Feature: Replace dict to enum types for globe-emoji option (with a choice on all enums)
+- Feature: Install script 'helloworld' with poetry script
+- Feature: Add logging support (for debug logs)
+- CI: Activate workflow on dev pushes
+- Core: Support for python 3.6.9
+- Core: Cruft - Update with template